# openapi.yaml
openapi: 3.1.0
info:
  title: HighFive! API
  version: 2.0.0
  description: >-
    Contrat backend v2, derive de `src/domain/**` (schemas zod), `src/api/*.ts`
    (appels reels du front) et `src/mocks/handlers/**` (regles metier
    appliquees cote mock). Voir `docs/v2/backend/SPEC.md` pour le detail des
    regles metier, du modele de persistance propose et des ecarts connus.
    Les schemas de composants sont generes depuis `src/domain/index.ts` par
    `pnpm export:schemas` (zod 4, `z.toJSONSchema`) dans `docs/v2/backend/schemas/`.
servers:
  - url: /api
    description: Prefixe fixe (voir `src/api/client.ts`), au-dessus de `VITE_API_URL`.
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Connexion
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginInput" }
      responses:
        "200":
          description: Session ouverte.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Identifiants invalides ou compte `deleted`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Inscription
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterInput" }
      responses:
        "201":
          description: Compte cree, session ouverte.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400":
          description: Corps invalide, ou email/pseudo deja pris (mock, `validation_error`) — recommande `409 conflict` cote backend, voir SPEC.md.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "409": { $ref: "#/components/responses/Conflict" }
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Deconnexion (invalide le jeton)
      responses:
        "204": { description: Jeton invalide. }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Session courante
      responses:
        "200":
          description: Personne connectee.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CurrentUser" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      operationId: updateCurrentUser
      tags: [users]
      summary: Mettre a jour son profil
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserProfileUpdateInput" }
      responses:
        "200":
          description: Profil mis a jour.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CurrentUser" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: R-P1 — compte `suspended` (bio/avatar verrouilles).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
  /auth/password-reset-request:
    post:
      operationId: requestPasswordReset
      tags: [auth]
      summary: Demander un lien de reinitialisation
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordResetRequestInput" }
      responses:
        "204":
          { description: "Toujours 204 : ne revele jamais si l'email existe." }
        "400": { $ref: "#/components/responses/BadRequest" }
  /auth/password-reset:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reinitialiser le mot de passe avec un jeton
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordResetInput" }
      responses:
        "204": { description: Mot de passe change. }
        "400": { $ref: "#/components/responses/BadRequest" }

  # ---------------------------------------------------------------- Tags
...
  /projects/{slug}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Fiche d'un projet
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      responses:
        "200":
          description: Projet complet.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "403":
          description: Projet prive et lecteur hors equipe (R-V3).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Modifier titre/accroche/description/tags/besoins/visibilite/participation
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectUpdateInput" }
      responses:
        "200":
          description: Projet mis a jour.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Reserve porteur/co-porteur (R-PR1).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Supprimer definitivement (R-PR7, confirmation nominative)
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectDeleteInput" }
      responses:
        "204":
          {
            description: "Projet supprime (differe 30 jours cote donnees, R-X3 ; voir SPEC.md).",
          }
        "400":
          description: Le titre saisi ne correspond pas (`confirmTitle`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Reserve au porteur et a l'administration.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /projects/{slug}/highfive:
    post:
      operationId: giveHighfive
      tags: [highfives]
      summary: Donner un highfive (R-H1, idempotent)
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      responses:
        "200":
          description: Etat et compteur recalcules cote serveur (R-X2).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HighfiveToggleResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: R-H2 — le porteur ne peut pas highfiver son propre projet.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: withdrawHighfive
      tags: [highfives]
      summary: Retirer un highfive (R-H4, sans notification)
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      responses:
        "200":
          description: Etat et compteur recalcules.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HighfiveToggleResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /projects/{slug}/highfives:
    get:
      operationId: listProjectHighfivers
      tags: [highfives]
      summary: Liste publique des personnes ayant highfive (R-H3)
      security: []
      parameters:
        - $ref: "#/components/parameters/SlugParam"
        - $ref: "#/components/parameters/CursorParam"
      responses:
        "200":
          description: Page de personnes.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedUserSummary" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------------------------------------------------- Memberships
...
  /projects/{slug}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: Commentaires (R-C1/R-V5 — masque les `hidden`, auteur resolu)
      security: []
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      responses:
        "200":
          description: Commentaires visibles.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CommentWithAuthor" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      operationId: createComment
      tags: [comments]
      summary: Commenter (R-C2 connecte et non suspendu, R-C4 un seul niveau)
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CommentCreateInput" }
      responses:
        "201":
          description: Commentaire publie.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comment" }
        "400":
          description: "Corps invalide, ou `parentId` pointant un commentaire deja reponse (R-C4)."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Compte suspendu (R-C2).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "404": { $ref: "#/components/responses/NotFound" }
  /comments/{commentId}/hide:
    post:
      operationId: hideComment
      tags: [comments]
      summary: Masquer un commentaire (porteur/co-porteur/admin, R-C3)
      parameters:
        - $ref: "#/components/parameters/CommentIdParam"
      responses:
        "204": { description: Commentaire masque. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Supprimer un commentaire (admin uniquement, R-C3)
      parameters:
        - $ref: "#/components/parameters/CommentIdParam"
      responses:
        "204": { description: Commentaire supprime. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Reserve a l'administration.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "404": { $ref: "#/components/responses/NotFound" }

  # --------------------------------------------------------------- Tasks
...
  /projects/{slug}/files:
    get:
      operationId: listFiles
      tags: [files]
      summary: Fichiers d'un projet (R-F4 — publics si le projet est public)
      security: []
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      responses:
        "200":
          description: Fichiers.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProjectFile" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      operationId: uploadFile
      tags: [files]
      summary: Deposer un fichier (R-F1 20 Mo/fichier, 200 Mo/projet ; R-F2 types autorises)
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
              required: [file]
      responses:
        "201":
          description: Fichier depose.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectFile" }
        "400":
          description: "Fichier manquant, type refuse (executable, R-F2), ou quota depasse (R-F1)."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "413":
          description: Fichier trop volumineux (alternative a un 400, au choix du backend — voir SPEC.md).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
  /files/{fileId}:
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Supprimer un fichier (R-F3 — deposant, porteur ou co-porteur)
      parameters:
        - $ref: "#/components/parameters/FileIdParam"
      responses:
        "204": { description: Fichier supprime. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ------------------------------------------------------------- Messagerie
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Jeton opaque (pas necessairement un JWT dans le mock : `nextId()`).
        Envoye `Authorization: Bearer <token>` par `src/api/client.ts`.
  parameters:
    CursorParam:
      name: cursor
      in: query
      required: false
      schema: { type: string }
      description: Curseur opaque de page (issu de `nextCursor`, `null` = derniere page).
...
    SlugParam:
      name: slug
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Slug" }
...
    CommentIdParam:
      name: commentId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Id" }
...
    FileIdParam:
      name: fileId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Id" }
...
  responses:
    BadRequest:
      description: "`validation_error` — corps ou parametres invalides."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorBody" }
    Unauthorized:
      description: "`unauthorized` — jeton absent ou invalide."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorBody" }
    Forbidden:
      description: "`forbidden` — role insuffisant pour cette action."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorBody" }
    NotFound:
      description: "`not_found` — ressource introuvable."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorBody" }
    Conflict:
      description: "`conflict` — etat deja atteint ou contrainte d'unicite (ex. email/pseudo deja pris)."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorBody" }
    UnprocessableEntity:
      description: "`validation_error` — corps syntaxiquement valide mais violant une regle metier (ex. R-PR1)."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorBody" }
  schemas:
...
    ApiErrorBody: { $ref: "./schemas/ApiErrorBody.json" }
...
    Comment: { $ref: "./schemas/Comment.json" }
    CommentCreateInput: { $ref: "./schemas/CommentCreateInput.json" }
...
    CurrentUser: { $ref: "./schemas/CurrentUser.json" }
...
    Id: { $ref: "./schemas/Id.json" }
...
    LoginInput: { $ref: "./schemas/LoginInput.json" }
...
    PaginatedUserSummary: { $ref: "./schemas/PaginatedUserSummary.json" }
...
    PasswordResetInput: { $ref: "./schemas/PasswordResetInput.json" }
    PasswordResetRequestInput:
      { $ref: "./schemas/PasswordResetRequestInput.json" }
...
    Project: { $ref: "./schemas/Project.json" }
...
    ProjectFile: { $ref: "./schemas/ProjectFile.json" }
...
    ProjectUpdateInput: { $ref: "./schemas/ProjectUpdateInput.json" }
    RegisterInput: { $ref: "./schemas/RegisterInput.json" }
...
    Session: { $ref: "./schemas/Session.json" }
    Slug: { $ref: "./schemas/Slug.json" }
...
    UserProfileUpdateInput: { $ref: "./schemas/UserProfileUpdateInput.json" }
    UserSummary: { $ref: "./schemas/UserSummary.json" }
...
    CommentWithAuthor:
      description: "Commentaire + auteur resolu (`commentWithAuthorSchema`, src/api/comments.ts)."
      allOf:
        - { $ref: "#/components/schemas/Comment" }
        - type: object
          properties:
            author: { $ref: "#/components/schemas/UserSummary" }
          required: [author]
...
    HighfiveToggleResponse:
      description: "R-H1/R-X2 (`highfiveToggleResponseSchema`, src/api/highfives.ts)."
      type: object
      properties:
        given: { type: boolean }
        highfiveCount: { type: integer, minimum: 0 }
      required: [given, highfiveCount]
      additionalProperties: false
...
    ProjectDeleteInput:
      description: "R-PR7 — confirmation nominative (schema anonyme, src/api/projects.ts)."
      type: object
      properties:
        confirmTitle:
          type: string
          description: Doit correspondre exactement au titre courant du projet.
      required: [confirmTitle]
      additionalProperties: false
...